fix: change the IRM interface

+++ test/mocks/IRMMock.sol
@@ -5,8 +5,8 @@ import "../../src/interfaces/IIRM.sol";
 
 contract IRMMock is IIRM {
     uint256 internal constant SECONDS_PER_YEAR = 365.2425 * 86400; // Gregorian calendar
-    int96 internal constant MAX_ALLOWED_INTEREST_RATE = int96(int256(uint256(5 * 1e27) / SECONDS_PER_YEAR)); // 500% APR
-    int96 internal constant MIN_ALLOWED_INTEREST_RATE = 0;
+    uint96 internal constant MAX_ALLOWED_INTEREST_RATE = uint96(uint256(5 * 1e27) / SECONDS_PER_YEAR); // 500% APR
+    uint96 internal constant MIN_ALLOWED_INTEREST_RATE = 0;
 
     uint256 internal interestRate;
 
@@ -14,8 +14,8 @@ contract IRMMock is IIRM {
         interestRate = _interestRate;
     }
 
-    function computeInterestRate(address market, address asset, uint32 utilisation) external returns (int96) {
-        int96 rate = computeInterestRateImpl(market, asset, utilisation);
+    function computeInterestRate(address market, address asset, uint32 utilisation) external returns (uint96) {
+        uint96 rate = computeInterestRateImpl(market, asset, utilisation);
 
         if (rate > MAX_ALLOWED_INTEREST_RATE) {
             rate = MAX_ALLOWED_INTEREST_RATE;
@@ -26,8 +26,8 @@ contract IRMMock is IIRM {
         return rate;
     }
 
-    function computeInterestRateImpl(address, address, uint32) internal virtual returns (int96) {
-        return int96(int256(uint256((1e27 * interestRate) / 100) / (86400 * 365))); // not SECONDS_PER_YEAR to avoid
+    function computeInterestRateImpl(address, address, uint32) internal virtual returns (uint96) {
+        return uint96(uint256((1e27 * interestRate) / 100) / (86400 * 365)); // not SECONDS_PER_YEAR to avoid
             // breaking tests
     }
